UHN-MA-45 abstract timer functions
