+++ components/Timer/Timer.js
@@ -5,6 +5,7 @@ import { connect } from 'react-redux';
 import { increaseTime, decreaseTime, countdown, clearTime, resetTime } from '../../store/actions';
 import ProgressCircle from 'react-native-progress-circle';
 import SetTimeButton from "./SetTimeButton";
+import { convertSeconds, convertSecondsToMinutes } from '../../services/time-functions.service';
 
 class Timer extends Component {
 
@@ -43,19 +44,6 @@ class Timer extends Component {
     }
   };
 
-  convertSeconds = (seconds) => {
-    let second = Math.floor(seconds % 3600 % 60);
-    if (second <= 9) {
-      second = "0" + second;
-    }
-    return second;
-  };
-
-  convertSecondsToMinutes = (seconds) => {
-    let minute = Math.floor(seconds % 3600 / 60);
-    return minute;
-  };
-
   componentDidMount() {
     if (this.props.isUsing) {
       this.interval = setInterval(this.countdown, 1000);
@@ -81,7 +69,7 @@ class Timer extends Component {
       >
         <SetTimeButton changeTimeHandler={this.incrementTimer}>+15</SetTimeButton>
         <Text style={{ fontSize: 72, fontWeight: "bold", color: "#67686B" }}>
-          {this.convertSecondsToMinutes(timeRemaining)}:{this.convertSeconds(timeRemaining)}
+          {convertSecondsToMinutes(timeRemaining)}:{convertSeconds(timeRemaining)}
         </Text>
         <SetTimeButton changeTimeHandler={this.decrementTimer}>-15</SetTimeButton>
